+++ Robust.Shared/GameObjects/Components/Collidable/ILookupWorldBox2Component.cs
@@ -6,7 +6,5 @@ namespace Robust.Shared.GameObjects
     public interface ILookupWorldBox2Component
     {
         Box2 GetAABB(Transform transform);
-
-        Box2 GetWorldAABB(Vector2? worldPos = null, Angle? worldRot = null);
     }
 }
